+++ src/parsers/manifest/dash/common/parse_representations.ts
@@ -29,8 +29,8 @@ import {
   IScheme,
   IContentProtectionIntermediateRepresentation,
 } from "../node_parser_types";
+import { convertSupplementalCodecsToRFC6381 } from "./convert_supplemental_codecs";
 import { getWEBMHDRInformation } from "./get_hdr_information";
-import { convertSupplementalCodecsToRFC6381 } from "./indexes/utils";
 import parseRepresentationIndex, {
   IRepresentationIndexContext,
 } from "./parse_representation_index";
